% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{agent_query}
\alias{agent_query}
\title{agent_query}
\usage{
agent_query(resp, con = duckdbfs::cached_connection())
}
\arguments{
\item{resp}{the response from the chat agent.
This is assumed to follow the JSON structure set by the system prompt.}

\item{con}{a DBI connection to the database.}
}
\description{
Takes the raw response text and returns a lazy tibble connection to the
table (view) containing the response.
}
