% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{create_prompt}
\alias{create_prompt}
\title{create_prompt}
\usage{
create_prompt(tbl, additional_instructions = "")
}
\arguments{
\item{tbl}{a remote table connection}

\item{additional_instructions}{length-1 character string containing
additional instructions to the agent.}
}
\description{
Create a system prompt for the LLM agent.
}
\details{
This uses an a system prompt aimed specifically at open models that
do not natively understand tool-calling or structured data replies.
}
